+++ src/com/intellij/idea/plugin/hybris/common/HybrisConstants.kt
@@ -227,6 +227,7 @@ object HybrisConstants {
 
     const val DEFAULT_HOST_URL = "localhost"
     const val DEFAULT_SSL_PROTOCOL = "TLSv1.2"
+    const val DEFAULT_SESSION_COOKIE_NAME = "JSESSIONID"
 
     const val HTTP_PROTOCOL = "http://"
     const val HTTPS_PROTOCOL = "https://"