+++ src/com/intellij/idea/plugin/hybris/settings/RemoteConnectionSettings.kt
@@ -45,6 +45,7 @@ class RemoteConnectionSettings : BaseState(), Comparable<RemoteConnectionSetting
     var isSsl by property(true)
     var isWsl by property(false)
     var sslProtocol by string(HybrisConstants.DEFAULT_SSL_PROTOCOL)
+    var sessionCookieName by string(HybrisConstants.DEFAULT_SESSION_COOKIE_NAME)
 
     var solrWebroot by string("solr")
     var hacWebroot by string("")