+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Other
 - Disable failing JRebelPlugin during build of the Searchable Options [#1325](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1325)
+- Support configurable session cookie name for remote HAC connection [#1334](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1334)
 
 ## [2025.1.0]